feat!: refactoring of CasparCGState
Remove mechanism to apply commands to a state, detach the diffing mechanism from the main class, split the diffing mechanism into multiple files
